pyproject.toml
[project]
name = "accountstatementextractor"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.17,<0.4.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "markdown (>=3.8,<4.0)",
    "pypdf (>=5.5.0,<6.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
